# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "La Casa De Rozgaar — Module 2 API",
    "description": "User, Talent & Career Intelligence Backend Module for La Casa De Rozgaar platform. Handles candidate profiles, assessment engine with anti-cheat integrity signals, skill gap analysis, job matching, career simulation, learning recommendations, interview intelligence, research intelligence, employer talent discovery, and workforce intelligence.",
    "version": "1.0.0",
    "contact": {
      "name": "La Casa De Rozgaar Engineering"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from POST /auth/login or POST /auth/register"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/workforce/gaps/analyze": {
      "post": {
        "summary": "Run organizational workforce skill gap analysis",
        "tags": [
          "Workforce Intelligence"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "profileId"
                ],
                "properties": {
                  "profileId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organizational skill gaps with employee impact counts"
          }
        }
      }
    },
    "/workforce/gaps/recommendation": {
      "post": {
        "summary": "Get Hire vs Upskill recommendations for organizational gaps",
        "tags": [
          "Workforce Intelligence"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "profileId"
                ],
                "properties": {
                  "profileId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hire vs Upskill vs Hybrid actions per skill"
          }
        }
      }
    },
...
  }
}